requirements.txt
pywin32==223
aircv==1.4.6
numpy==1.15.1
Pillow==5.3.0
win32gui==221.6